Use a Monte Carlo simulation for n columns
This generates a set of possible letter for each of the first n columns and calculates how many possible passwords there are. It repeats this and updates the average for a more accurate guess until the user presses Ctrl+C to interrupt it. The current estimate is printed to screen once every second.

It took 37 seconds to run one million valid simulations, which estimated 3.47 possible passwords with 3 columns of information. However, I ran this given all 5 columns of information (taking 1 min 42 seconds for one million simulations) and the output was 1.54 passwords. This is wrong because it should be exactly one. The manual states "Only one combination of the available letters will match a password." I realize now I need a separate method for generating valid possible letter combinations.